+++ package.json
@@ -16,7 +16,7 @@
   },
   "dependencies": {
     "@expo/vector-icons": "^14.0.2",
-    "@react-native-async-storage/async-storage": "^2.0.0",
+    "@react-native-async-storage/async-storage": "1.23.1",
     "@react-native-community/slider": "^4.5.3",
     "@react-navigation/native": "^6.0.2",
     "expo": "~51.0.28",
